Initialization-logic-fix
if we call the constructor here, it will be too early because all @Input variables are undefined or empty now, if we call initializeSelection() at this moment, all values we take will be undefined. We call ngOnInit() , it will wait for all @Input created, then it binds data from the front-end tag, then it starts calling initializeSelection() to do whatever logic we have in there.