changing maximumError in angular.json
The maximumError is a rule that you set for the build process on a hosting service like Netlify we currently use for Github. It's like a checkmark to check the size of our project, let's say if maximumError is 2mb and our project is 2.2mb now, Netlify hosting service will throw an error.