Fixed logic for calculating slope of slider
Fixed the logic error in calculating the slope of the linear function of the slider. Before the fix, the code was incorrectly using the slider's initial Y coordinate stored in the Joint object, instead of its Y coordinate from the previous frame.  The fix was to consistently use both prevJointPosition.x and prevJointPosition.y when calculating n.