Fix wrong animation with mechanism when slider angle equals 90 degrees (vertical)
Before the fix, when the slider became vertical (angle = 90 degrees), the slider didn't move, and the coupler link was strangely animated by stretching out and in. After the fix, everything became normal as it should be. The issue was in the old code, which mistakenly used the old point coordinates instead of the current ones. In other words, when the crank moves an angle theta to a new position, we should use this new position to draw a circle analysis instead of the last position.